Completed bomSortCallback() and associated helper methods.
New UI elements:
BOM.bomContentLabels: a 7 by (BOM table height -1) 2D list of labels.
New helpers:
-BOM.createBomLabels(): Creates the contents of bomContentLabels for the correct table height.
-BOM.populateBomRow(): Sets the label texts for one row of bomContentLabels from a given bomPart object.
-BOM.attachBomRow(): Attaches one row of bomContentLabels to bomTable.

The callback currently deletes the previous contents of bomContentLabels and creates new labels.
This could be made more efficient in the future by preserving exsiting objects,
especially in the case of the new table being larger.
However, this will function as a prototype and is less error-prone.

Signed-off-by: Joe Baker <jbaker0428@gmail.com>