Digikey scrape now checks for out of stock items.
Instead of the "Quantity Availale" field on the page saying 0,
DK has a form field to request a user-inputted quantity of the product.
If scrape() finds a <form> here, it sets the product's quantity field to 0.
Swapping between C1 and C2 on the BOM page now updates the product info pane properly with this fix.