Partially fixed find_similar_parts attribute checks.
The change to the Part constructor in this commit solves an issue where the Part instances in the find_similar_parts()
results lists had empty attributes dictionaries.

The changes to product_updater are necessary despite performance losses in order to ensure that find_similar_parts functions correctly,
as it needs the DB version of the part instance to be up to date.

The find_similar_parts queries are still not 100% correct.
Given a part with non-null attributes, the query will return parts with no attributes.
In other words, it is not satisfying the condition that returned parts must have all of the attributes in self.attributes.

Signed-off-by: Joe Baker <jbaker0428@gmail.com>